update comment and creation comment -> movie note

+++ src/main/java/com/usach/movie_backend/movies/domain/Movie.java
@@ -34,7 +34,7 @@ public class Movie {
     @Column(name = "views")
     private Integer views;
     @Column(name = "note")
-    private Float note;
+    private Double note;
     @Column(name = "id_director")
     private Integer idDirector;
     @Column(name = "id_producer")
